requirements.txt
streamlit==1.32.0
pandas==2.2.1
selenium==4.15.2
webdriver-manager==4.0.2
python-dotenv==1.0.0
pyautogen[openai]==0.8.5
openai==1.12.0
Pillow==9.5.0
tiktoken==0.5.2
asyncio==3.4.3
pytesseract==0.3.13
beautifulsoup4==4.12.3
requests==2.31.0
python-dateutil==2.8.2
numpy==1.26.4